+++ bodo/pandas/_util.cpp
@@ -1,4 +1,4 @@
-#include "_duckdb_util.h"
+#include "_util.h"
 
 std::variant<int8_t, int16_t, int32_t, int64_t, uint8_t, uint16_t, uint32_t,
              uint64_t, bool, std::string, float, double, arrow::TimestampScalar>
